Part 2. Overriding the fq_class_name variable
Comparing the `handleTemplatedMixins` and `handleRegularMixins` methods,
we can notice that in the `handleRegularMixins` method the `fq_class_name`
variable is overridden if the sought method is found in the mixin.

Whereas in the `handleTemplatedMixins` method this does not happen.
And I think therein lies the error.

Since the `class_storage` variable can be overridden in the
`handleTemplatedMixins` method, the `fq_class_name` variable should
also be overridden.